applied J. Russell Smyth 's patches
removed nullable and isNullable attributes added by Steven Caswell - see
previous message - in short, these are not DB attributes, but JDBC
attributes, unreliable, db dependent, and are handled by the already
existing "required" attribute. 

removed charOctetLength - this is db implementation details that should
never be reflected in the db independent model. (I will be sending
another suggestion on how to handle this later) 

changed scale,precisionRadix, and ordinalPosition  from Integer to int -
no good reason to keep this as an Integer, and makes their use more
cumbersome. 

typeCode is an int and is expected to be a java.sql.Types value.
Though have preserved the String type property.

removed getIsTextType - dosnt belong here. It is already in TypeMap
where it does belong.

removed getTypeString - this is not the right place to create SQL code -
it could vary by platform. Explained more fully in previous message.

removed getSQLTypeCode - dosnt belong here, it is in TypeMap already.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/sql/trunk@147468 13f79535-47bb-0310-9956-ffa450edef68